requirements.txt
Django>=1.11,<1.12
djangorestframework==3.6.3
